# example-catalog-management-service/src/main/resources/openapi/catalog-api.yaml
openapi: 3.0.3
info:
  title: Catalog Management Service API
  description: >
    E-commerce product catalog management. Resources form a hierarchy —
    Catalog → Category (self-referencing tree) → Product — which is exactly the
    shape attribute-based access control (ABAC) is designed for. This baseline
    has no authentication; auth is layered on in later phases.
  version: 0.1.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: http://localhost:8080
    description: Local

tags:
  - name: Catalog
    description: Top-level catalogs
  - name: Category
    description: Hierarchical categories within a catalog
  - name: Product
    description: Products within a category

paths:
  # ---------------------------------------------------------------- Catalogs
  /api/v1/catalogs:
    get:
      tags: [Catalog]
      summary: List catalogs
      operationId: listCatalogs
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: A page of catalogs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Catalog]
      summary: Create a catalog
      operationId: createCatalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catalog'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/v1/catalogs/{catalogId}:
    parameters:
      - $ref: '#/components/parameters/CatalogId'
    get:
      tags: [Catalog]
      summary: Get a catalog
      operationId: getCatalog
      responses:
        '200':
          description: Catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catalog'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Catalog]
      summary: Update a catalog
      operationId: updateCatalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catalog'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Catalog]
      summary: Delete a catalog
      operationId: deleteCatalog
      responses:
        '204':
          description: Deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # -------------------------------------------------------------- Categories
  /api/v1/catalogs/{catalogId}/categories:
    parameters:
      - $ref: '#/components/parameters/CatalogId'
    get:
      tags: [Category]
      summary: List categories in a catalog
      operationId: listCategories
      parameters:
        - name: parentId
          in: query
          description: When set, returns only direct children of this category.
          required: false
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: A page of categories (the rows the caller is authorized to see)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Category]
      summary: Create a category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  /api/v1/catalogs/{catalogId}/categories/{categoryId}:
    parameters:
      - $ref: '#/components/parameters/CatalogId'
      - $ref: '#/components/parameters/CategoryId'
    get:
      tags: [Category]
      summary: Get a category
      operationId: getCategory
      responses:
        '200':
          description: Category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Category]
      summary: Update a category
      operationId: updateCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Category]
      summary: Delete a category
      operationId: deleteCategory
      responses:
        '204':
          description: Deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------- Products
  /api/v1/catalogs/{catalogId}/categories/{categoryId}/products:
    parameters:
      - $ref: '#/components/parameters/CatalogId'
      - $ref: '#/components/parameters/CategoryId'
    get:
      tags: [Product]
      summary: List products in a category
      operationId: listProducts
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: A page of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Product]
      summary: Create a product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/catalogs/{catalogId}/categories/{categoryId}/products/{productId}:
    parameters:
      - $ref: '#/components/parameters/CatalogId'
      - $ref: '#/components/parameters/CategoryId'
      - $ref: '#/components/parameters/ProductId'
    get:
      tags: [Product]
      summary: Get a product
      operationId: getProduct
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Product]
      summary: Update a product
      operationId: updateProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Product]
      summary: Delete a product
      operationId: deleteProduct
      responses:
        '204':
          description: Deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    CatalogId:
      name: catalogId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      description: 0-based page index. A bounds violation is a 400 VALIDATION_FAILED (no clamping).
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    PerPage:
      name: perPage
      in: query
      description: Page size (1–100). A bounds violation is a 400 VALIDATION_FAILED (no clamping).
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: Access denied
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    UnprocessableEntity:
      description: A domain rule rejected the request (e.g. an illegal tag value)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    ServiceUnavailable:
      description: A required dependency was unavailable (fail-closed; the request was rejected)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'

  schemas:
    PageEnvelope:
      type: object
      description: >
        The shared list envelope. Every public list returns a page of items plus the exact total —
        and under ABAC that total is subject-relative: count is the number of rows THE CALLER is
        authorized to see, across all pages (never items.length). page/perPage echo the request.
        Rows are ordered createdAt ASC, id ASC (a fixed total order; clients do not choose the sort).
      required: [count, page, perPage]
      properties:
        count:
          type: integer
          format: int64
          description: Total rows the caller is authorized to see, across all pages.
        page:
          type: integer
          minimum: 0
          description: 0-based page index, echoing the request.
        perPage:
          type: integer
          minimum: 1
          maximum: 100
          description: Page size, echoing the request.

    CatalogPage:
      allOf:
        - $ref: '#/components/schemas/PageEnvelope'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Catalog'

    CategoryPage:
      allOf:
        - $ref: '#/components/schemas/PageEnvelope'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Category'

    ProductPage:
      allOf:
        - $ref: '#/components/schemas/PageEnvelope'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Product'

    Catalog:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true

    CatalogRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000

    Category:
      type: object
      required: [id, catalogId, name]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        catalogId:
          type: string
          format: uuid
          readOnly: true
        parentId:
          type: string
          format: uuid
          nullable: true
          description: Parent category id; null for a root category.
        name:
          type: string
        description:
          type: string
        tags:
          type: object
          additionalProperties: true
          description: >
            Dictionary-validated tags on this category. Each value is a string (single-valued key) or
            an array of strings (multi-valued key), per the tag definition.

    CategoryRequest:
      type: object
      required: [name]
      properties:
        parentId:
          type: string
          format: uuid
          nullable: true
          description: Parent category id; omit/null for a root category.
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        tags:
          type: object
          additionalProperties: true
          description: >
            Tags to assign, validated against the tag dictionary. Each value is a string (single-valued
            key) or an array of strings (multi-valued key). An unknown key or illegal value is rejected
            with 422; nothing is silently dropped.

    Product:
      type: object
      required: [id, categoryId, name, priceCents, currency]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        categoryId:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        description:
          type: string
        sku:
          type: string
        priceCents:
          type: integer
          format: int64
          description: Price in minor units (e.g. cents).
        currency:
          type: string
          description: ISO 4217 currency code.
          example: USD

    ProductRequest:
      type: object
      required: [name, priceCents, currency]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        sku:
          type: string
          maxLength: 64
        priceCents:
          type: integer
          format: int64
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3

    ProblemDetail:
      type: object
      description: >
        An RFC-7807 problem object (application/problem+json). Five standard members plus two documented
        extensions (errorCode, timestamp). A client branches on the typed errorCode, not the human detail.
      required: [status, errorCode]
      properties:
        type:
          type: string
          description: A stable, relative, opaque identifier for the problem kind (e.g. /problems/tag-value-illegal). Not dereferenced.
        title:
          type: string
          description: A short, status-stable summary of the problem kind.
        status:
          type: integer
          format: int32
          description: The HTTP status code.
        detail:
          type: string
          description: A human, instance-specific explanation.
        instance:
          type: string
          description: The request path that produced the error.
        errorCode:
          type: string
          description: The machine-stable code a consumer branches on. The union of codes this service can emit.
          enum:
            - ACCESS_DENIED
            - RESOURCE_NOT_FOUND
            - VALIDATION_FAILED
            - TAG_VALUE_ILLEGAL
            - DEPENDENCY_UNAVAILABLE
        timestamp:
          type: string
          format: date-time
          description: When the error was produced (correlation).